requirements.txt
pytube
moviepy
requests
python-dotenv
pydub
tenacity
rich